Fixed bug with adding removed files

+++ oxen-rust/src/lib/src/core/v0_19_0/rm.rs
@@ -23,6 +23,7 @@ use crate::constants::STAGED_DIR;
 use crate::constants::VERSIONS_DIR;
 use crate::model::Commit;
 use crate::model::StagedEntryStatus;
+use crate::model::merkle_tree::node::DirNode;
 
 use rmp_serde::Serializer;
 use serde::Serialize;
@@ -74,7 +75,6 @@ fn remove(
         data_type_counts: HashMap::new(),
     };
 
-    // TODO: Right now, this will delete the file even if oxen rm fails. Is that an issue?
     for path in paths {
         // Remove dirs
         if path.is_dir() {
@@ -90,7 +90,7 @@ fn remove(
             }
 
             // Remove files
-        } else {
+        } else if path.is_file() {
             match remove_file(repo, &maybe_head_commit, path) {
                 Ok(entry) => {
                     if let Some(entry) = entry {
@@ -116,6 +116,24 @@ fn remove(
             log::debug!("REMOVING FILE: {full_path:?}");
             if full_path.exists() {
                 util::fs::remove_file(&full_path)?;
+            }
+        } else {
+            let mut maybe_dir_node = None;
+            log::debug!("Found non-existant path: {path:?}");
+            if let Some(ref head_commit) = maybe_head_commit {
+                let path = util::fs::path_relative_to_dir(path, &repo.path)?;
+                let parent_path = path.parent().unwrap_or(Path::new(""));
+                maybe_dir_node = CommitMerkleTree::dir_with_children(repo, head_commit, parent_path)?;
+            }
+
+            if let Ok(Some(_dir_node)) = get_dir_node(&maybe_dir_node, path) {
+                log::debug!("non-existant path {path:?} was dir. Calling remove_dir"); 
+                remove_dir(repo, &maybe_head_commit, path.to_path_buf());
+
+            } else if let Ok(Some(_file_node)) = get_file_node(&maybe_dir_node, &path) {
+                log::debug!("non-existant path {path:?} was file. Calling remove_file");
+                let opts = RmOpts::from_path(&path);
+                remove_file(repo, &maybe_head_commit, &path);
             }
         }
 
@@ -227,6 +245,7 @@ pub fn remove_file(
     maybe_head_commit: &Option<Commit>,
     path: &Path,
 ) -> Result<Option<StagedMerkleTreeNode>, OxenError> {
+    println!("Remove file");
     let repo_path = repo.path.clone();
     let versions_path = util::fs::oxen_hidden_dir(&repo.path)
         .join(VERSIONS_DIR)
@@ -577,6 +596,25 @@ fn get_file_node(
         if let Some(node) = node.get_by_path(path)? {
             if let EMerkleTreeNode::File(file_node) = &node.node {
                 Ok(Some(file_node.clone()))
+            } else {
+                Ok(None)
+            }
+        } else {
+            Ok(None)
+        }
+    } else {
+        Ok(None)
+    }
+}
+
+fn get_dir_node(
+    dir_node: &Option<MerkleTreeNode>,
+    path: impl AsRef<Path>,
+) -> Result<Option<DirNode>, OxenError> {
+    if let Some(node) = dir_node {
+        if let Some(node) = node.get_by_path(path)? {
+            if let EMerkleTreeNode::Directory(dir_node) = &node.node {
+                Ok(Some(dir_node.clone()))
             } else {
                 Ok(None)
             }
